Improve performance
The 'base64' module is not used in this script, so it can be ignored.
If the key file already exists, the script first creates the new key, and then reads the existing key from the key file. So here, if the key doesn't really exist, the script can create it.
You can read the key in one line of code from the key file (better readability).