Improve readability
Instead of reading a fixed number as the length of bytes in the donkey file, we can instantly calculate the number of bytes in the donkey file and then use it for donkey.
This has two advantages. First, it helps to understand how the script works, and second, the user can simply enter his own file path instead of the donkey string, without the need for additional manual calculations.